Move serializeQueryParams to its own file so that consuming applications and addons can import it directly
